# backend/src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0",
        "description": "Documentation for the API endpoints"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "summary": "Registrar um novo usuário",
                "description": "Cria uma nova conta de usuário no sistema",
                "tags": ["Auth"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário criado com sucesso."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "409": {
                        "$ref": "#/components/responses/Conflict"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/auth/sign-in": {
            "post": {
                "summary": "Autenticar usuário",
                "description": "Autentica um usuário e retorna os dados de autenticação",
                "tags": ["Auth"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autenticação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/rag/chat": {
            "post": {
                "summary": "Gerar resposta para mensagem",
                "description": "Gera uma resposta baseada na mensagem enviada pelo usuário",
                "tags": ["RAG"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta gerada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Esta é uma resposta gerada pelo sistema."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/rag/insight": {
            "post": {
                "summary": "Gerar insight",
                "description": "Gera um insight baseado nos dados fornecidos",
                "tags": ["RAG"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InsightRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Insight gerado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "insight": {
                                            "type": "string",
                                            "example": "Este é um insight gerado pelo sistema."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/rag/upload-documents/": {
            "post": {
                "summary": "Fazer upload de documentos",
                "description": "Faz upload de documentos para o sistema",
                "tags": ["RAG"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "files": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Arquivos salvos com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Files saved!"
                                        },
                                        "files": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "file1.pdf",
                                                "file2.docx"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/rag/collection/{collectionName}": {
            "get": {
                "summary": "Obter arquivos salvos",
                "description": "Obtém os arquivos salvos em uma coleção específica",
                "tags": ["RAG"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "collectionName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Nome da coleção"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivos obtidos com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "files": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "file1.pdf",
                                                "file2.docx"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "delete": {
                "summary": "Excluir documento",
                "description": "Exclui um documento de uma coleção específica",
                "tags": ["RAG"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "collectionName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Nome da coleção"
                    },
                    {
                        "name": "fileName",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Nome do arquivo a ser excluído"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Documento excluído com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Document with filename \"example.pdf\" deleted successfully."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/message/": {
            "get": {
                "summary": "Obter mensagens do usuário",
                "description": "Obtém as mensagens associadas ao usuário autenticado",
                "tags": ["Message"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensagens obtidas com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Message"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/message/count-assistant": {
            "get": {
                "summary": "Contar mensagens do assistente",
                "description": "Conta o número de mensagens do assistente para o usuário autenticado",
                "tags": ["Message"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contagem obtida com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "integer",
                                            "example": 42
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/user/": {
            "get": {
                "summary": "Obter dados do usuário",
                "description": "Obtém os dados do usuário autenticado",
                "tags": ["User"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do usuário obtidos com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "RegisterRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "joao@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "senha123"
                    }
                },
                "required": ["name", "email", "password"]
            },
            "LoginRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "joao@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "senha123"
                    }
                },
                "required": ["email", "password"]
            },
            "AuthResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "joao@example.com"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                }
            },
            "ChatRequest": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Como posso ajudar você hoje?"
                    }
                },
                "required": ["message"]
            },
            "InsightRequest": {
                "type": "object",
                "properties": {
                    "insight": {
                        "type": "string",
                        "example": "Dados para gerar insight"
                    }
                },
                "required": ["insight"]
            },
            "Message": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "content": {
                        "type": "string",
                        "example": "Olá, como posso ajudar?"
                    },
                    "owner": {
                        "type": "string",
                        "enum": ["user", "assistant"],
                        "example": "assistant"
                    },
                    "userId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                }
            },
            "ValidationErrorItem": {
                "type": "object",
                "properties": {
                    "field": {
                        "type": "string",
                        "example": "email"
                    },
                    "message": {
                        "type": "string",
                        "example": "Email inválido"
                    }
                }
            }
        },
        "responses": {
            "BadRequest": {
                "description": "Requisição inválida",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Senha incorreta."
                                }
                            }
                        }
                    }
                }
            },
            "ValidationError": {
                "description": "Erro de validação dos dados enviados",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ValidationErrorItem"
                                    },
                                    "example": [
                                        {
                                            "field": "email",
                                            "message": "Email inválido"
                                        },
                                        {
                                            "field": "password",
                                            "message": "A senha deve ter pelo menos 6 caracteres"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Recurso não encontrado",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Usuário não encontrado."
                                }
                            }
                        }
                    }
                }
            },
            "Conflict": {
                "description": "Conflito",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "O email desse usuario ja foi cadastrado"
                                }
                            }
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Não autorizado",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Token inválido ou expirado."
                                }
                            }
                        }
                    }
                }
            },
            "InternalServerError": {
                "description": "Erro interno do servidor",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Internal Server Error."
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}